Design: 채팅 UI 수정
