+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_API_KEY=http://43.202.191.111:8080
