+++ src/components/Header/Header.css
@@ -11,7 +11,7 @@
 }
 
 .logo {
-    height: 50px; /* 로고의 높이 조정 */
+    height: 30px; /* 로고의 높이 조정 */
 }
 
 .nav ul {
