+++ src/components/Header/Header.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import './Header.css';
+import logo from '../../assets/images/logo.png';
 
 const Header = () => {
     return (
         <header className="header">
             <div className="logo-container">
-                <img src="/images/logo.png" alt="Logo" className="logo" />
+                <img src={logo} alt="Logo" className="logo" />
             </div>
             <nav className="nav">
                 <ul>
